pyproject.toml
[tool.poetry]
name = "open-autonomy"
version = "0.21.18"
description = "Open Autonomy Framework"
authors = ["developer-valory <develope@valory.xyz>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "autonomy"}, {include = "packages"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
# Core runtime
open-aea = { version = "==2.2.1", extras = ["all"] }
# CLI deps. `click`, `pytest`, `coverage` come transitively via
# `open-aea[all]`; only the IPFS CLI plugin is autonomy-specific.
open-aea-cli-ipfs = "==2.2.1"
# Optional plugins — declared via extras below.
# [hwi] and [docker] are intentionally kept out of [all]:
#   * HWI's transitive chain (ledgerwallet -> hidapi + Pillow) has
#     no armv7 wheels, breaking multi-platform Docker image builds.
#   * `docker` is only needed for docker-compose deployments; users
#     deploying to Kubernetes or localhost don't need it.
open-aea-ledger-ethereum = {version = "==2.2.1", optional = true, extras = []}
open-aea-ledger-ethereum-hwi = {version = "==2.2.1", optional = true}
docker = {version = "==7.1.0", optional = true}

# ---------------------------------------------------------------------------
# Direct runtime imports from `packages/valory/**` that aren't covered
# by the `open-aea[all]` / ledger-plugin transitive chain. Each entry
# is directly `import`-ed by at least one runtime module in this repo
# (checked via `ast`-walk of `autonomy/` and `packages/valory/` outside
# `tests/` / `test_tools/`).
#
# Version policy:
#   * Unpinned (`*`) when the effective range is already constrained
#     by a ledger plugin or `open-aea[all]` transitively — we trust
#     the upstream to dictate a compatible version. Matches the `{}`
#     entries in the package YAMLs.
#   * Pinned / bounded only when the dep is version-sensitive for
#     correctness. Such ranges must stay in sync with the
#     corresponding package YAML.
# ---------------------------------------------------------------------------
open-aea-ledger-cosmos = "==2.2.1"
# open-aea-ledger-solana = "==2.2.1"  # NOTE: can't coexist with our
# dev tooling. `open-aea-ledger-solana` pulls `anchorpy<0.19` which
# transitively requires `cachetools<5`; that clashes with
# `tomte[tox]`'s `cachetools>=7.0.1` (via `tox 4.46`). Re-enabling
# needs either the anchorpy upstream to loosen, or moving solana
# into an opt-in extra.
web3 = "*"
requests = "*"
grpcio = "==1.78.0"
aiohttp = ">=3.8.5,<4.0.0"
certifi = "*"
multidict = "*"
asn1crypto = ">=1.4.0,<1.5.0"
ecdsa = "*"
py-ecc = ">=8,<10"

# ---------------------------------------------------------------------------
# TODO(cleanup): entries below are declared here only to satisfy
# `tox -e check-dependencies`. That tool reads
# `[tool.poetry.dependencies]` but:
#   * ignores entries with `optional = true` unless they also carry
#     an `extras` key (so `open-aea-ledger-ethereum` above is
#     invisible; we restate it here without `optional`);
#   * doesn't read the dev group (so `pytest-asyncio` is invisible);
#   * flags `protobuf` as missing even though it arrives transitively
#     via `open-aea[all]`.
# The authoritative distribution manifest is `setup.py`, so these
# redundant entries are harmless. The right long-term fix is to
# patch the checker to understand `optional` + dev-group deps.
protobuf = "*"
pytest-asyncio = "*"

[tool.poetry.extras]
cli = [
    "open-aea-cli-ipfs",
]
chain = [
    "open-aea-ledger-ethereum",
]
docker = [
    "docker",
]
hwi = [
    "open-aea-ledger-ethereum-hwi",
]
all = [
    "open-aea-cli-ipfs",
    "open-aea-ledger-ethereum",
]

[tool.poetry.scripts]
autonomy = "autonomy.cli:cli"

[tool.poetry.group.dev.dependencies]
tomte = { extras = ["tox", "docs"], version = "==0.6.5" }
mkdocs-monorepo-plugin = "^1.0.5"
open-aea-ci-helpers = "==2.2.1"
open-aea-dev-helpers = "==2.2.1"
# Local CI/release tooling — used by `make generators`, `tox -e
# fix-doc-hashes`, etc. Not imported by any runtime code.
open-aea-helpers = { path = "plugins/aea-helpers", develop = true }
# Test tooling: `open-aea-test-autonomy` (our own plugin) supplies the
# base test classes imported from `packages/valory/**/test_tools/`;
# `hypothesis`, `pytest`, `pytest-asyncio` are only used by test
# modules (pytest is guarded via a `_PytestMissing` stub in
# `test_tools/` so skill imports work without pytest installed).
open-aea-test-autonomy = { path = "plugins/aea-test-autonomy", develop = true }
hypothesis = "*"
pytest = "*"
pytest-asyncio = "*"
# Used by the docker-compose tendermint sidecar
# (`deployments/Dockerfiles/tendermint/app.py`) and a legacy test helper
# (`packages/valory/agents/register_reset/tests/helpers/slow_tendermint_server/app.py`).
# Kept out of the main deps because the runtime autonomy wheel uses
# `autonomy/deploy/_http_server.py` (stdlib-only Flask-compatible shim)
# instead.
Flask = ">=3.1.0,<4.0.0"
Werkzeug = ">=3.1.0,<4.0.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
